feat: update mungectl API
